+++ domain/src/main/java/com/dida/domain/model/nav/post/Comments.kt
@@ -1,9 +1,9 @@
 package com.dida.domain.model.nav.post
 
 data class Comments(
-    val commentId: Int,
-    val postId: Int,
-    val userId: Long,
+    val commentId: Long,
+    val postId: Long,
+    val userId: Int,
     val content: String,
     val userName: String,
     val userImgUrl: String,
