+++ feature/community-detail/src/main/java/com/dida/community_detail/DetailCommunityNavigationAction.kt
@@ -5,6 +5,6 @@ sealed class DetailCommunityNavigationAction {
     class NavigateToCommentMore(val commentId: Long): DetailCommunityNavigationAction()
     object NavigateToCommunityMore: DetailCommunityNavigationAction()
     class NavigateToUpdateCommunity(val postId: Long): DetailCommunityNavigationAction()
-    class NavigateToUserProfile(val userId: Long): DetailCommunityNavigationAction()
+    class NavigateToUserProfile(val userId: Int): DetailCommunityNavigationAction()
     class NavigateToCardDetail(val cardId: Long): DetailCommunityNavigationAction()
 }
