+++ feature/recent-nft/src/main/java/com/dida/recent_nft/RecentNftNavigationAction.kt
@@ -1,6 +1,6 @@
 package com.dida.recent_nft
 
 sealed class RecentNftNavigationAction {
-    class NavigateToRecentNftItem(val nftId: Int): RecentNftNavigationAction()
+    class NavigateToRecentNftItem(val nftId: Long): RecentNftNavigationAction()
     object NavigateToCardRefresh: RecentNftNavigationAction()
 }
