+++ model/src/main/java/com/dida/model/data/response/GetMainResponse.kt
@@ -10,7 +10,7 @@ data class GetMainResponse(
 )
 
 data class HotItemResponse(
-    @SerializedName("cardId") val cardId: Int,
+    @SerializedName("cardId") val cardId: Long,
     @SerializedName("imgUrl") val imgUrl: String,
     @SerializedName("name") val name: String,
     @SerializedName("price") val price: String,
@@ -25,7 +25,7 @@ data class HotSellerResponse(
 )
 
 data class RecentCardResponse(
-    @SerializedName("cardId") val cardId: Int,
+    @SerializedName("cardId") val cardId: Long,
     @SerializedName("cardName") val cardName: String,
     @SerializedName("userName") val userName: String,
     @SerializedName("imgUrl") val imgUrl: String,
